view, insert, update and delete projects
ProjectsController-Mustafa-V1.0
This script is used to view, insert, update or delete a project, It uses
the CRUD Model functions to deal with the database.
To deal with this script all you need is to create a file for each
function and just do the two steps below:
- first step >> require the file from current path
- second step >> call the required function and it will do all the
needed work
- congrats you are done
The script was tested and it is working well.

-- Note: There is some bugs in the projects table in the database [TBD]